templates/groups/users permission check added

+++ apps/backend/src/templates/resolvers/mutation/delete-template.ts
@@ -1,10 +1,13 @@
 import { MutationResolvers } from '../../../common/types/gql.g';
 import { prisma } from '../../../app';
+import { checkUserTemplateCriticalPermission } from '../../services';
 
 export const deleteTemplate: MutationResolvers['deleteTemplate'] = async (
   parent,
-  { templateId }
+  { templateId },
+  { user }
 ) => {
+  await checkUserTemplateCriticalPermission(user.id, templateId);
   await prisma.template.delete({ where: { id: templateId } });
   return true;
 };
